+++ esibd/plugins_internal/Pressure/Pressure.py
@@ -154,6 +154,7 @@ class PressureController(DeviceController):
                         # self.print(f'Read pressure for channel {c.name}', flag=PRINT.DEBUG)
                     except Exception as e:
                         self.print(f'Failed to parse pressure from {msg}: {e}', PRINT.ERROR)
+                        self.errorCount += 1
                         self.pressures[i] = np.nan
                 elif channel._controller == channel.TPG and self.tpgInitialized:
                     msg = self.TPGWriteRead(message=f'PR{channel.id}', lock_acquired=True)
@@ -167,6 +168,7 @@ class PressureController(DeviceController):
                             self.pressures[i] = np.nan
                     except Exception as e:
                         self.print(f'Failed to parse pressure from {msg}: {e}', PRINT.ERROR)
+                        self.errorCount += 1
                         self.pressures[i] = np.nan
                 else:
                     self.pressures[i] = np.nan