+++ esibd/plugins_internal/temperature/Temperature.py
@@ -157,6 +157,7 @@ class TemperatureController(DeviceController):
                 self.temperatures[i] = float(value)
             except ValueError as e:
                 self.print(f'Error while reading temp: {e}', PRINT.ERROR)
+                self.errorCount += 1
                 self.temperatures[i] = np.nan
 
         # toggle cryo on off to stabilize at temperatures above what is possible with minimal power